+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Available as `edge`
 
 - TBD
 
+## [1.15.3] - 2026-08-20
+
+- Fixed: Restored build output when using the BuildKit builder [Issue-2461](https://github.com/caprover/caprover/issues/2461)
+- Fixed: Tar file deployments failing from the dashboard [Issue-2463](https://github.com/caprover/caprover/issues/2463)
+
 ## [1.15.2] - 2026-08-14
 
 - Fixed: Legacy custom Captain NGINX templates failing after upgrading to v1.15.0 [Issue-2455](https://github.com/caprover/caprover/issues/2455)